requirements.txt
# Django e extensões
Django
djangorestframework
django-cors-headers

# WebSockets
channels
channels-redis
daphne

# Redis
redis

# HTTP requests
requests

# WebSocket client (para robôs)
websockets

# Utilitários
python-dotenv

# Desenvolvimento e debug (opcional)
django-debug-toolbar

# Produção (opcional)
gunicorn
psycopg2-binary # Para PostgreSQL
whitenoise      # Para servir arquivos estáticos

